Bec Braughton Additions
I quashed as many bugs as I was able; the validator said a lot of dashes in the comment was not good, so I did remove that... I don't think I did 'fancy' indenting, I did fix some out of order indenting and cleared white space, hope that is ok; it's best coding practice. I also added some animations on the Fb and Twitter divs.